+++ ptn/missionalertbot/constants.py
@@ -80,6 +80,7 @@ PROD_ROLEAPPS_CHANNEL = 867820665515147293 # role-applications on the live serve
 PROD_UPVOTE_EMOJI = 828287733227192403 # upvote emoji on live server
 PROD_O7_EMOJI = 806138784294371368 # o7 emoji on live server
 PROD_FC_COMPLETE_EMOJI = 878216234653605968 # fc_complete emoji
+PROD_DISCORD_EMOJI = 1122605426844905503 # Discord emoji on live
 PROD_HAULER_ROLE = 875313960834965544 # hauler role ID on live server
 PROD_WINELOADER_ROLE = 881809680765165578 # wine loader role on live
 PROD_CC_ROLE = 869340261057196072 # CC role on live server
@@ -131,6 +132,7 @@ TEST_ROLEAPPS_CHANNEL = 1121736676247609394 # role-applications on the test serv
 TEST_UPVOTE_EMOJI = 849388681382068225 # upvote emoji on test server
 TEST_O7_EMOJI = 903744117144698950 # o7 emoji on test server
 TEST_FC_COMPLETE_EMOJI = 884673510067286076 # fc_complete emoji
+TEST_DISCORD_EMOJI = 1122605718198026300 # Discord emoji on live
 TEST_HAULER_ROLE = 875439909102575647 # hauler role ID on test server
 TEST_WINELOADER_ROLE = 1119189364522623068 # wine loader role on test
 TEST_CC_ROLE = 877220476827619399 # CC role on test server
@@ -357,6 +359,9 @@ def o7_emoji():
 def fc_complete_emoji():
   return PROD_FC_COMPLETE_EMOJI if _production else TEST_FC_COMPLETE_EMOJI
 
+def discord_emoji():
+  return PROD_DISCORD_EMOJI if _production else TEST_DISCORD_EMOJI
+
 def hauler_role():
   return PROD_HAULER_ROLE if _production else TEST_HAULER_ROLE
 
@@ -423,7 +428,7 @@ def seconds_short():
 def seconds_long():
   return PROD_SECONDS_LONG if _production else TEST_SECONDS_LONG
 
-def training_category():
+def training_cat():
   return PROD_TRAINING_CATEGORY if _production else TEST_TRAINING_CATEGORY
 
 def training_mission_command_channel():